fix: inject MUI styles before Tailwind
The CSS injected last gets higher priority, and as Tailwind is built on
top of MUI, it only makes sense to give Tailwind a higher priority.

A similar fix was previously added in a commit from February 13, 2024
and subsequently removed on June 23, leaving me to wonder why I was not
able to override MUI styles.